Refactor in preparation for Lua Part 2
Should be final part
