+++ src/dungeon_data.c
@@ -24,6 +24,7 @@
 #include "config_terrain.h"
 #include "game_legacy.h"
 #include "player_instances.h"
+#include "gui_soundmsgs.h"
 #include "post_inc.h"
 
 /******************************************************************************/
@@ -204,6 +205,35 @@ TbBool player_has_heart(PlayerNumber plyr_idx)
 {
     return thing_exists(get_player_soul_container(plyr_idx));
 }
+
+void add_heart_health(PlayerNumber plyr_idx,HitPoints healthdelta,TbBool warn_on_damage)
+{
+    struct Thing* heartng = get_player_soul_container(plyr_idx);
+    if (!thing_is_invalid(heartng))
+    {
+        struct ObjectConfigStats* objst = get_object_model_stats(heartng->model);
+        long old_health = heartng->health;
+        long long new_health = heartng->health + healthdelta;
+        if (new_health > objst->health)
+        {
+            SCRIPTDBG(7,"Player %u's calculated heart health (%I64d) is greater than maximum: %ld", heartng->owner, new_health, objst->health);
+            new_health = objst->health;
+        }
+        heartng->health = new_health;
+        if (warn_on_damage)
+        {
+            if (heartng->health < old_health)
+            {
+                event_create_event_or_update_nearby_existing_event(heartng->mappos.x.val, heartng->mappos.y.val, EvKind_HeartAttacked, heartng->owner, heartng->index);
+                if (is_my_player_number(heartng->owner))
+                {
+                    output_message(SMsg_HeartUnderAttack, 400);
+                }
+            }
+        }
+    }
+}
+
 
 /** Returns if given dungeon contains a room of given kind.
  *
