+++ src/main.cpp
@@ -2699,6 +2699,7 @@ void update(void)
         event_process_events();
         update_all_events();
         process_level_script();
+        process_fx_lines();
         if ((game.numfield_D & GNFldD_Unkn04) != 0)
             process_computer_players2();
         process_players();
@@ -4331,6 +4332,8 @@ LONG __stdcall Vex_handler(
         return EXCEPTION_CONTINUE_EXECUTION; // Thrown by OutputDebugStringW, intended for debugger
     } else if (exception_code == DBG_PRINTEXCEPTION_C) {
         return EXCEPTION_CONTINUE_EXECUTION; // Thrown by OutputDebugStringA, intended for debugger
+    }else if (exception_code == 0xe24c4a02) {
+        return EXCEPTION_EXECUTE_HANDLER; //Thrown by luaJIT for some reason
     }
     LbJustLog("Exception 0x%08lx thrown: %s\n", exception_code, exception_name(exception_code));
     return EXCEPTION_CONTINUE_SEARCH;
